requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
anyio==4.13.0
auto-py-to-exe==2.50.0
av==17.0.0
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
ctranslate2==4.7.1
customtkinter==5.2.2
darkdetect==0.8.0
Eel==0.18.2
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.4
onnxruntime==1.24.4
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
protobuf==7.34.1
PyAudio==0.2.14
pycparser==3.0
Pygments==2.20.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
pynput==1.8.1
pyparsing==3.3.2
pyperclip==1.11.0
pystray==0.19.5
pywin32-ctypes==0.2.3
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.1
rich==14.3.3
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
SpeechRecognition==3.15.2
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
zope.event==6.1
zope.interface==8.2
